+++ src/features/mypage/actions/add-mypage.ts
@@ -2,9 +2,9 @@
 import { sendLineNotifyMessage } from "@/apis/line-notify/send-message";
 import { createMypage } from "@/apis/prisma/fetch-mypage";
 import { ERROR_MESSAGES, SUCCESS_MESSAGES } from "@/constants";
+import { validateMypage } from "@/features/mypage/utils/validate-mypage";
 import { formatCreateContentMessage } from "@/lib/format-for-line";
 import type { MypageContext } from "../stores/mypage-context";
-import { validateMypage } from "../utils/validate";
 
 type AddMypageState = ServerAction & {
 	data?: MypageContext;
