+++ src/features/mypage/utils/validate-mypage.ts
@@ -1,0 +1,14 @@
+import { FORM_ERROR_MESSAGES } from "@/constants";
+import { mypageSchema } from "../schemas/mypage-schema";
+
+export function validateMypage(formData: FormData) {
+	const mypageValidatedFields = mypageSchema.safeParse({
+		title: formData.get("title"),
+		quote: formData.get("quote"),
+		url: formData.get("url"),
+	});
+	if (!mypageValidatedFields.success)
+		throw new Error(FORM_ERROR_MESSAGES.INVALID_FORMAT);
+
+	return mypageValidatedFields;
+}
